Initial Commit
Setting up a new repository for this collection of things.  I have been
writing a lot of these to expand functionality of my
notifications/statusbar for better quick information on the status of my
system and the various features it has.  This is an effort to formalize
all of the various bits of unified code and *semi* useful pythonic
wrappers I have written around system information or helper binaries.