[acpi][icons] ACPI update and icon font change
Two changes in one :sigh:

Update the logic in the ACPI handler to be a bit more useful:
- Slower update rate to avoid higher churn
- State tracking to avoid spamming updates
- Attempt at keeping a permanent notification in the drawer of the
  current battery level (transient and timeout based)

Switch the font used for some things.  This is the general material
design font and provides more icons and consistently sized icons for
stuff like bluetooth.  Will probably add more to other plugins in the
future as I explore the options.