[mpris] Add initial mpris handler
This attempts to use `playerctl` and some weird asyncio subprocesses to
keep monitoring active mpris players and update the HUD and
notifications on changes.  Still very much WIP, issues with broken state
due to playerctl not reporting changes correctly or mishandling the
updates causes old players to remain in a 'playing' state.